feat: improve extension of default html elements
